:sparkles: Save for ltr help that returned nothing
Saved in DB the word asked that returned nothing. By using the attribute `valid`.

Admin can list the non `valid` by using the command `!helpUnvalid`.
Also, using the `!helpAdd` on a non `valid` word will not trigger an error. It will just update the definition, and validate the word

Close #3

Signed-off-by: Maxime Cattet <maximecattet@gmx.com>